Textures: Allow compressed texture from PBO
Apex Legends is using a pixel unpack buffer to populate a
compressed 3D texture.  We had an explicit check in validation
that prevented it, possibly due to emulated formats.

Rather than prevent it altogether, check whether the format
is emulated and proceed if not.

Also add a test that exposes the problem.

Test: PBOCompressedTexture3DTest
Bug: b/236478448
Bug: angleproject:7464
Change-Id: I90105a396bf0a465c62d67910aad2819ba7f8831
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3719161
Commit-Queue: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>