Expose EXT_compressed_ETC1_RGB8_sub_texture on Metal and GL
Updated the test to ensure that this extension
is exposed on Apple GPUs when using Metal.

Updated emulatedEtc1 limitation to hide this extension
when the main ETC1 extension is hidden.

Real WebGL apps cannot enable this extension because
it is not exposed in WebIDL.

Removed bogus entry from IsETC1Format and use the helper
consistently in all validation functions.

Simplified GetNativeCompressedFormat.

Bug: angleproject:7471
Change-Id: I61321fadad7d962358d0fefecd08aaddaedd2ec2
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3737762
Reviewed-by: Jonah Ryan-Davis <jonahr@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>