Vulkan: Fix image arguments passed to functions in GLSL
It's impossible to provide the format qualifier to an image function
parameter in GLSL, because `layout()` is not acceptable there.  SPIR-V
doesn't accept formatless image parameters without the
StorageImageReadWithoutFormat and StorageImageWriteWithoutFormat
capabilities, which are not universally available.

Instead, this change monomorphizes functions with image parameters.
Note that this was already previously done for R32F images for emulation
purposes.

Bug: angleproject:7484
Change-Id: Id287c4a93b32deb3fd8be982c32fa58fd2c848d8
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3763074
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>