Add flag --save-screenshots (implied by --screenshot-dir).
--screenshot-dir and --render-test-output-dir behave differently on
Android. --screenshot-dir is passed as is and therefore means a device
directory, and --render-test-output-dir has special handling in Android
wrappers that substitutes this flag with a temporary location on the
device then copying files to the local directory set by this flag.

Cody pointed out that there are many docs referring to --screenshot-dir
so behavior should be preserved. So --save-screenshots is implied when
this flag is set. Otherwise, it defaults to false, so
--render-test-output-dir just on its own won't save screenshots, only
when --save-screenshots is added as well.

This also allows to produce other artifacts without necessarily
producing screenshots by using --render-test-output-dir without
--save-screenshots.

Bug: angleproject:7299
Change-Id: Ib1a268ddeb7eba9f120e3f37a429ea0e4bbe2411
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3785296
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Jamie Madill <jmadill@chromium.org>
Commit-Queue: Roman Lavrov <romanl@google.com>