Context: Also disable ASTC on mState when emulation is disabled.
GL_KHR_texture_compression_astc_ldr was incorrectly exposed on the
GL_EXTENSIONS string in cases where no native support was available and
the emulation was not built.

To prevent this the context must disable it on mState.mExtensions in
addition to mSupportedExtensions.

Bug: angleproject:7415
Change-Id: I1161e460b0b224a03557053b5804c104e270f388
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3788402
Reviewed-by: Jamie Madill <jmadill@chromium.org>
Commit-Queue: Lubosz Sarnecki <lubosz.sarnecki@collabora.com>