Vulkan: Optimize transform feedback buffer tracking
Prior to this CL, if transform feedback was active at the time of render
pass closure, its buffers were cached in ContextVk.  Later, these
buffers were used to close the render pass if they were used for any
other reason (such as vertex attribute).

However, this meant that the render pass could close unnecessarily if
transform feedback was ended right after the render pass is closed.  The
closure of the render pass was an awkward place to cache the used
transform feedback buffers (because at that point, the buffers are
actually no longer used).

Instead, this change makes sure that the buffers are cached when
transform feedback buffers are first used by the render pass, and the
cache is cleared at the end of the render pass.

Bug: angleproject:4622
Change-Id: I31c0a1e20d48f2e261e2cf37adb0a46db683e6fb
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3788309
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Auto-Submit: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>