Metal: Stop pre-creating the Provoking Vertex CmdBuffer
The provoking vertex command buffer was being created
regardless if whether or not it was used which is left
over from when the backend used to pre-enqueue command
buffers. That's no longer true so we don't need to pre-create
this command buffer.

I think this adds a 5% perf increase to the MotionMark
"Leaves" test.

Bug: angleproject:7619
Change-Id: Ic921911eeef0f063ea6eba8ba7c307157243b938
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3863739
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Gregg Tavares <gman@chromium.org>