Make BitSetT::Mask return a bit set
This allows a Mask function to be implemented for BitSetArray, which is
added in this CL.  When using larger bitsets on 32-bit systems, the
current Mask implementation prohibits its use.

This is in preparation for a follow up change that uses Mask on such a
bitset.

Bug: angleproject:7369
Change-Id: If995d96ec1583a546f20bff277f3223e2f2490f5
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4031493
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>