Support versioned library names
Previously ANGLE_GLESV2_LIBRARY_NAME was expected to be the library name
without extension, and on Linux .so would be appended. This didn't allow
for building ANGLE as libEGL.so.1 and libGLESv2.so.2 as is common.

This change checks if the library name already contains a dot, and if so
doesn't append the platform specific extension suffix.

Bug: b/259282560
Change-Id: I81cda218180ae7fe6321e1ac7e4eff7257bf1493
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4038450
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Nicolas Capens <nicolascapens@google.com>
Reviewed-by: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Auto-Submit: Nicolas Capens <nicolascapens@google.com>