Make sure D3D11 textures never lose bind flags
We hit an issue where a texture was used alternatively as a UAV and a
render target. When used as a render target, ensureBindFlags created a
new storage with just the render target bind flag. When used as a UAV,
ensureBindFlags created a new storage again, this time with just the UAV
bind flag. And so on.

This change updates ensureBindFlags to ensure all previous bind flags
are preserved when creating a new storage.

Bug: angleproject:7855
Change-Id: I8c03c011e3df6eee5fbad22e45663e8fee731d57
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4112763
Reviewed-by: Stephen White <senorblanco@chromium.org>
Commit-Queue: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Jamie Madill <jmadill@chromium.org>