GLES1: Optimize texcoord varyings
Use vec3 instead of vec4.  Either 2 or 3 components are used based on 2D
or cube map textures.  On hardware that can pack the varyings, this can
reduce pressure on varyings.

Additionally, this is a minor optimization to texture coord calculation
(when !draw_texture) as one row of the 4x4 matrix multiplication doesn't
need calculation.

Bug: angleproject:6644
Change-Id: I2abac511a0ced575c089dfb3a3b506f51a111a54
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4116676
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>