Capture/Replay: Deal with swap called in different contexts
Move recording the context setup of the main context to the
same location where all other context setups are recorded and
drop the assertion that checks that swap is always called from
the same context.

Invoke the context setup at the time the secondary contexts are set
up, and make sure the main context is correctly mapped.

Bug: angleproject:7911
Change-Id: I327bce318b1a0e26ffdbf096343f99cedd78c116
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4236541
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Commit-Queue: Cody Northrop <cnorthrop@google.com>