Fix EGLImage importing problem
Chrome DrDC uses EGLImage to share textures between contexts.
Sometimes, chrome will create texture with a GLES 3.x context,
and then create EGLImage from it, and then import EGLImage to
a GLES 2.0 (WebGL 1.0) context. In this case, ANGLE may report
errors if the original texture is created with formats which
are not supported with GLES 2.0, but actually the texture can
be sampled and rendered by ANGLE. Fix the problem by adding
mIsTexturable & mIsRenderable in Image, and calculating them
against the original context which owns the texture.

Bug: chromium:1417294
Change-Id: I4c304422f0d0fc01a48c942d05758f8c2ce438e4
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4265252
Commit-Queue: Peng Huang <penghuang@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>