Copy d3dcompiler_47.dll on ARM64 Windows builds
Previously d3dcompiler_47.dll was not copied because this DLL is an
inbox component on ARM64 Windows. However the inbox is not necessarily
updated frequently enough, so we should copy (and ship) this DLL
on ARM64 Windows as well as on other Windows architectures.

Bug: chromium:1416385, chromium:1406077
Change-Id: I463f215012c2b48f11c8f40e104c22622afa6b2a
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4285717
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>