Add support for permissive pixel comparison
Adds support for running ANGLE pixel tests with a more permissive
inexact matching algorithm. This is done by passing in
--use-permissive-pixel-comparison=1 to the test runner.

This is intended to reduce the amount of manual triage work
required on CLs that are expected to have a larger amount of
differences, such as SwiftShader rolls.

On the bots, this will normally be disabled, but will be enabled
if "Use-Permissive-Angle-Pixel-Comparison: True" is present as
a CL footer.

This footer is not yet included automatically anywhere, so
there should be no functional change as a result of this CL yet.

Bug: angleproject:7985
Change-Id: Ie815fac42edb2198dd4d115fc50650504df136c0
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4288612
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>