Vulkan: Always checkCompletedCommands from finishResourceUse
There is one difference in crrev.com/c/4244823, that before the CL, we
always check every command to see if it finished or not. After the CL,
we stop as soon as we find ResourceUse is finished. This CL gets back
old behavior to help speed up garbage collection.

Bug: chromium:1420669
Change-Id: I6bf833bd3347ee683b6b9d78628ac9618845c9e2
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4304802
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>
Commit-Queue: Charlie Lao <cclao@google.com>