Makes ToposortStructs() return consistent results
This patch introduces a step in ToposortStructs() that
sorts the nodes by name. Doing this ensures consistent
results between calls. Without this, the order depends
upon the pointer value, which varies depending upon when
called.
Ensuring consistent results for the same input is important
for other parts of this code that use the string as key.

Bug: chromium:1423136
Change-Id: I19f053df5cf86c81b800056e86b742d8c34af7d4
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4357255
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>