Tests: Add Minecraft Bedrock trace
This CL adds a new Minecraft trace.  The app has changed engines since
our first trace, updating to the Bedrock engine.

Also, our first trace was recorded on Qualcomm, and the trace does not
behave correctly on ARM.

This also bumps the original minecraft version, but the content is
unchanged.

Test: angle_trace_tests --gtest_filter="*minecraft_bedrock*"
Bug: b/183647942
Bug: b/197965659
Change-Id: Id65889c65df8a6a3a1560ccbe4fce239982d98de
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4376790
Reviewed-by: Roman Lavrov <romanl@google.com>