Avoid calling GetGpuIDFromDisplayID in Mac GPU info collection.
To detect the currently active GPU, CGLQueryRendererInfo is used.
Calling any CGL function loads the system's OpenGL driver even
when we don't intend on using OpenGL for rendering.

Confirmed that the OpenGL framework is not loaded when Chrome is run
with Metal on a single-GPU system.

Bug: chromium:1433697
Change-Id: I824e7663948169123ad8c87d6b3867616233c329
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4428302
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Jonah Ryan-Davis <jonahr@google.com>