Improve adb root handling
If `id -u` returns 0, adbd is already running as root (convenient fast
path for local dev and seems we have that on bots too)
If `which su` returns something, we should be able to get root.
If "su" is there but we can't get adb root, raise.

Go back to up to 10 seconds wait.
On my local device adb root is routinely taking ~4 seconds.
adb root restarts adbd on the device so it's expected not to be fast.

Bug: b/276742336
Change-Id: If720b9176ba18ebd9734c0b8288eabc5970b5ef8
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4508392
Commit-Queue: Roman Lavrov <romanl@google.com>
Reviewed-by: Cody Northrop <cnorthrop@google.com>