Don't use system info collection in the GL backend on Android
The only way to gather system info on Android is with Vulkan or an
existing OpenGL context (which we have), so skip the call to
angle::GetSystemInfo on Android and use the OpenGL strings that
are already gathered.

It also tends to crash when creating Vulkan instances on some devices.

Make sure the ARM vendor ID is detected from the renderer strings.

Bug: chromium:1447111
Change-Id: Ic5c4109b422ca240bf5e8fc5cef6a2dae6d7d29d
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4562933
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>