Metal: Put a limit on the total MTLLibrary objects cached
Use an MRU for holding the cached libraries in mtl::LibraryCache.

Using a maximum cache size of 128 and a GC size of 32 yields the same
ratio of cache hits as before (40%) while cycling out the more unique
shaders that are not re-used. Lowering the cache size to 64 or 96
yields a hit rate of 30-35%.

Bug: chromium:1329376
Change-Id: I75dd6b20cf3f8b67c4b3c31da4e05e1857ad61db
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4614362
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>