Remove _shardN suffix from tests when it's not necessary.
My understanding is that we only need this for --split-shard-samples.
Otherwise we still get _shardN suffix from the shard but the test only
ran on a single shard so there is no point in that.

_shardN added in https://crrev.com/c/3937044

--split-shard-samples disabled in https://crrev.com/c/4605335

Example of _shardN:
https://ci.chromium.org/ui/p/angle/builders/ci/linux-nvidia-gtx1660-perf/1704/overview

Bug: angleproject:7671
Change-Id: Iaec4eb81e6ebad424cfc2ebfcbc7eda312cd1bba
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4615538
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Roman Lavrov <romanl@google.com>