gl: Do not propagate copyTexSubImage2D error to client
https://chromium-review.googlesource.com/c/angle/angle/+/4662393 ensured
that BlitGL::copyTexSubImage allowed the caller to fallback when the
underlying copyTexSubImage2D call failed, but the GL error was also
propagated to the client due to the Context::handleError call in
CheckError - that error still causes webgl conformance failures even
though the pixel expectations are now satisfied.

This CL makes BlitGL::copyTexSubImage use getError directly so that the
error is not propagated to the client.

Bug: chromium:1458040
Change-Id: I3e4c06871c55d449facbe471c78e84f68b529f60
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4671325
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>