[Android] Update NDK version check for hardware buffer.
The function used for checking if hardware buffer is supported requires
NDK 29, but is currently checking if 26+ is in use.

This was not an issue before, but when updating the minimum SDK (and
NDK) version in Chromium, this led to an inconsistency.

For now, increasing the version requirement to 29+, and therefore also
removing the unnecessary branch for lock planes support.

Bug: chromium:1497004
Change-Id: I3d6e60e51db0831ee81d817aec90760d7680cc31
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4985624
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>