Consolidate parallel compilation in front-end
This cleans up the multiple compilation task implementations in the
backends and consolidates them in the front-end.  The front-end is then
able to do the compilation in an unlocked tail call instead if
desired (in a future change).

This change is in preparation for having the program link tasks directly
wait on the shader compilation tasks.  As a result, the "shader resolve"
should not be needed to access the shader compilation results; it should
be enough to wait for the compilation job.  This change therefore moves
post-processing of results to the compilation job itself as they did not
need to actually be done after compilation is done (merely after
translation is done).  As a side effect, shader substition and other
debug features should now work for the GL backend as they are now done
before back-end compilation.

Bug: angleproject:8297
Change-Id: Ib9274b1149fadca7545956a864d6635b6cba5c3a
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/4994655
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>