FrameCapture: Handle context Reset with no calls
The code was assuming that setDefaultResetCalls adds
reset calls for an entry point, but it has logic for
skipping if the current state is already default.

This CL checks to see if we added any calls for the
entry point before trying to iterate through them.

Test: Flip capture
Bug: b/308208067
Change-Id: I8a214b0eb0df068d6b3172c0ca1406f6751f91ea
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5114271
Reviewed-by: Mike Schuchardt <mikes@lunarg.com>
Commit-Queue: Cody Northrop <cnorthrop@google.com>