deqp: Unused test expectations do not cause failure, only log
deqp is the only suite that appears to be handling unused expectations
this way.

This causes an inconvenience where a test cannot be pre-emptively added
to the expectations file. For example, a VK-GL-CTS roll might be adding
a new test that is failing in some configurations. The expectation
cannot be added to the expectations file pre-emptively because of
the failure due to the unused expectation, but that is the more
convenient way to solve this problem than patching the roll manually.

Bug: angleproject:8507
Change-Id: If2ee65d9119043c42e8ca0705a1cc8c40827fae3
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5238992
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Roman Lavrov <romanl@google.com>
Auto-Submit: Roman Lavrov <romanl@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>