Vulkan: Remove call to angle::GetSystemInfo()
This function is expensive, especially when it creates a temp VkInstance
to query device info.

It was only used to check if running on a Pixel device, but the
information retrieved was to make a workaround slightly more precise.
This change does away with that information and applies the workaround
more generally.

If the workaround precision is necessary, the necessary information
should be retrieved directly without making the expensive
angle::GetSystemInfo() call.

Bug: b/327509734
Change-Id: Id097b427c59efde16a5811d484823b207c1a8917
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5331602
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Auto-Submit: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Cody Northrop <cnorthrop@google.com>